Three files worked on AutoModeTesting, MotorServoClass(ReUseAbleTest), TeleOpTankMode.
MotorServoClass:
Is just the file ReUseAbleTest renamed to make more sense for what it does.

TeleOpTankMode:
Switched over using handlers to help with a delay that is needed on Y
button presses so servos do not flip so fast.

 AutoModeTest:
Nothing major done just switching over the call commands from the
old ReUseAbleTest ones to the new MotorServoClass commands
because of the rename beforehand.