Allterrain: Worked on cleaning up this program some rewriting motor names ect. A list of all new names shall be added at a later date, this will have all names and how to wire the motors to run correctly.
ColorTest: Added a lot of auto programing. The numbers just need to be worked on for how far the robot runs forward to be correct, but that should be all the needed major changes. Also is kind of split in to two programs with one part currently commented to out so that we can run it if the other part does not work. Most likely we will need numbers updated to match what the color sensor really sees not what we used to think it did.

     Enum_Test: Just messing around with stuff, please ignore for now will talk more about in a later update.