Meeting update done by Justin:
Worked on stuffz #YoloSwag420:

Worked on the values in ColorTest everything up to the first set of "If else" checks works correctly and is finished.

Small edits in allterrain to make it clearer and run better, turbo mode will be added soon and it renamed to "Comp TeleOp" which would be are finished program for teleop.