Meeting update done by Justin and bradley!
We worked on making code to dump the climbers from the robot holding spot in to the goal for them! We did this because the part of the team in charge of stratgey said that it was good to do this because it gets us 40 points! So we follow are orders and did are job well in the service of team beta. #Teamβ.

On a real note we finished the code to dump the climbers from the robot in the the goal behind the color beacon from being parked in front of it all in auto mode, run time is about 8-10 seconds for just this part.