12/20/16
Added code to OmniBot for a full range of functions for the drive control teleop.
Functions include all servos wheel drive and lifts and shooter.

Todo:
Create function in newTeleop or OmniBot to trigger a vuforia nav drive controller to take over and drive the robot in on a target so the drive does not need to do this.
Finish functions in newTeleop to have correct button on-off triggers for the lift and other servos.
Finish function for the shooting lift servo so that its on a set time.
Finish code for nav debugging, y seems to read x?

Create code to toggle forward drive on keypad.
    Look in to using the mode key for this

Finish code for gyro sensor, look in to possiable 300ms delay because of other sensor reads, also look in to temp disablement of the sensor for better read time.