12/28/16
Added testDiffrentFront as a clone of OmniBot with a few small code updates.

Updated code in omniBot to add functions for the dropper servo for the lift and functions to control it.

Added code to newTeleop to control the functions added in omniBot.

Updated static var's in autoBlue AutoOpMode autoRed to swich from a set var for vuforia stop distance to a const that is set to 5.08cm